Add support for a side navigation bar. Also fixes a bug where the highlighting of the chosen section in the top navigation was not showing up correctly.
With this support, it is assumed that side navigation is a subtree of the top navigation.
For example, if a url is base.com/foo/bar, then foo is the top nav location and bar
is the left nav location.

The tree can be specified in the config as follows. Note that Home currently cannot support
a side navigation bar.

nav:
  - item:
      name: Home
      url: /
  - item:
      name: Top Item 2
      url: /topitem2
      sidenav:
        - item:
            name: "sidenav1"
            url: "/sidenav2-1"
	- item:
            name: "sidenav2"
            url: "/sidenav2-2"
  - item:
      name: "Top Item 3"
      url: /topitem3
      sidenav:
        - item:
            name: "side nav sub dir"
	    dir_url: "/sidedir"
	    subgroup:
              - item:
                  name: "multibindings"
                  url: "/multibindings"
              - item:
                  name: "multibindings2"
                  url: "/multibindings2"

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=304258839